Added method ImagingReceiver.generateReceivedPmfForUniformInput .
This method computes the PMF of the received MIMO VLC signal, given that
the transmitted signal is uniformly distributed.

Note that this function was mostly written a while ago; I should check to
make sure that changes in progress are complete and that the documentation
matches.

Also, I've had some difficulty quantifying the expected error for this PMF
(I think that's what I was working on before taking a break on this.)