Removed output min_noise_to_delta_ratio from MIMOCapacity.generateReceivedPmfForUniformInput and updated tests in MIMOCapacityTest to not check for this output.
min_noise_to_delta_ratio was originally intended to be used as a measure
of the "quantization noise floor" artifact.  However, as explained on p.9
of lab book #4 (Imaging Receivers & Photodetector Arrays), this concept is
flawed.  I have since replaced this flawed concept with min and max bounds
on the calculated differential entropy.