Investigated the cause of calulating a negative mutual information when the gain is zero, where the negative mutual information was greater than 5 standard deviations (as calculated for variance_H) from zero.
To debug this, I modified MIMOCapacity.calcualteCamacityForUniformInput
to output h_y and pmf (local variables for this method) so that it can be
inspected in the calling method.

I also added method MIMOCapacity.calculateDIffEntropyOfClippedNormal to
calculate the differential entropy using a more precise (noise-free)
version of the PMF.

And I created a test method
MIMOCapacityTest.testCalculateCapacityForUniformInputeVarianceH to
invertigate this problem and to determine whether the calculated capacity
was within 5*sqrt(variance_H) of the correct capacity.  This test failed.

An explanation of this problem is in lab book #4 (Imaging Receivers &
Photodetector Arrays), p.36-38 and p.40.  Proposed solutions (to be
implemented) are on p.39 of lab book #4.