Added method MIMOCapacity.generateTransformedUniformPmf.
This method work determines which bins are reachable through Monte Carlo
simulations with uniform x; then sets the PMF for each of those bins to be
uniform so that their total probability adds to 1.

However, this method has a few flaws with respect to the computation time
that may be required as described on p. 43-44 of lab book #4 (Imaging
Receivers & Photodetector Arrays): if a very unlikely bin is hit once, the
method may take an extremely long time to finish.  This greatly varying
run time makes the method difficult to use.

TODO: instead, replace this with a deterministic algorithm.