Renamed MIMOCapacity.generateTransformedUniformPmf to MIMOCapacity.generateUniformPmfForGx.
Modified this function to accept the same parameters as
MIMOCapacity.computeUniformPmfForGx (in hopes that they may be used
interchangably) since they serve the same purpose.  However,
MIMOCapacity.generateUniformPmfForGx does require Gx_max to be close to
the upper bounds of its bin.

Added a check to MIMOCapacity.generateUniformPmfForGx to throw a warning
if Gx_max is too far from the upper bounds of its bin.