Removed the second while loop from MIMOCapacity.generateUnformPmfForGx, which kept simulating until the bin with the least hits has at least tc hits.
The reasoning is that if they bins are less likely than the bin containing
Gx_max, then it a PMF that approximates the probability of hitting those
bins as zero may be a better approximation than a PMF that approximates
the probability for those bins as equal to the probability of hitting
other reachable bins.  These less likely bins may occur when the reachable
values of G*x glancingly touch a corner (or a negligibly small part) of a
bin.

Also move the test to check if li_Gx_max was successfully calculated to
immediately after it is calculated.  And changed it to throw an error if
li_Gx_max cannot be successfully calculated.

TODO: This method still needs to be tested.