Debugged the checkCorollary1 failures in SmithCapacity.computeCapacityOnlyAmplitudeCon.  Simplified SmithCapacity.checkCorollary1.
The problem (of determining that n is not optimal for a particular A, even
when it still is) seems to arise from precision error in the "optimal" poi
and voi returned by fmincon.  This problem was resolved by improving the
precision of fmincon.  I also do a minor correction of the returned
"optimal" distribution by pushing the two extreme points of increase to -A
and A.

SmithCapacity.checkCorollary1 was rewritten to better handle precision
errors and to be much simpler in structure.