Commit the work that was in progress (on using h(Ax)=h(x)+log|det(A)| and the Gram-Schmidt QR-decomposition to get a transform for y to more efficiently compute the UnifX capcity lower bound) when I started writing my dissertation (which interrupted this work).
Making a commit because I don't remember exactly what was in progress and
I want to be able to go back to this state if necessary.

It seems like I had just modified MIMOCapacity.computeCapacityLBUnifX to
try using
MIMOCapacityLBUnifX.approximateUnifXLBForNegligibleNoiseAndFullColRankG
(falling back to the slower original method if G is not full column rank).

To test this, I wrote the unit test
MIMOCapacityLBUnifXTest.testApproximateUnifXLBForNegligibleNoiseAndFullColRankG3x2.

And I had modified the QR-decomposition in
MIMOCapacityLBUnifX.approximateUnifXLBForNegligibleNoiseAndFullColRankG to
use the Gram-Schmidt method (instead of the default for MATLAB's qr
method) so that we get a positive A=Q'*G as required by our new test.