Adjusted the max_nbins_linprog code in MIMOCapacity.computeCapacityLBUnifGx.
Changed the value from 6e2 to 6e4 as described in the comment; however, I think it may become necessary to either change this (to adapt to changing number of dimensions) or remove this code as it currently guarantees neither the quality of the PMF nor the computation time.

Replaced the Gx_max code with MIMOCapacity.computeUExtremes.